Move literal, rows and errors to their own files
Keeps sql smaller and easier to scan. Moves constants to files where
they are easy to open by name.
